requirements.txt
djangorestframework==3.15.2
psycopg2==2.9.10
Pillow==11.0.0